Update Syntax and separate modules.
