+++ modules/elements.js
@@ -1,0 +1,9 @@
+/* Form Items */
+const form = document.querySelector('form');
+const bookList = document.getElementById('bookList');
+
+/* Navigation Bar Links */
+const menuItems = Array.from(document.querySelectorAll('.menu-item'));
+const contentItems = Array.from(document.querySelectorAll('.content'));
+
+export { form, bookList, menuItems, contentItems };
