+++ modules/objects.js
@@ -1,0 +1,7 @@
+import { Books } from './classes.js';
+
+/* Books Collection Object */
+const books = new Books();
+books.displayBooks();
+
+export { books };
